Id Number Command done and scrreng vaidation done

+++ app/Http/Controllers/TrainingSessionController.php
@@ -16,6 +16,7 @@ use App\Models\TrainingCenterBasedPermission;
 use App\Models\TrainingCenterCapacity;
 use App\Models\TrainingMaster;
 use App\Models\TrainingMasterPlacement;
+use App\Models\TrainingPlacement;
 use App\Models\TrainingSession;
 use App\Models\TrainingSessionTraining;
 use App\Models\TraininingCenter;
@@ -723,17 +724,20 @@ class TrainingSessionController extends Controller
     }
     public function resetScreen($training_session_id)
     {
-        foreach (Volunteer::where('training_session_id', $training_session_id)->get() as $volunteer) {
 
-            foreach (Status::all() as $status)
+        if (count(TrainingPlacement::where(['training_session_id' => request()->route('training_session')])->get()) > 0) {
 
-                if ($volunteer->id == $status->volunteer_id) {
-
-                    Status::find($status->id)->update(['acceptance_status' => 1]);
+            return redirect()->back()->withErrors('Reseting Screening Is Not Allowed Because Training Placement is Already Done!!  Reset Training Placement To do this Task ');
+        } else {
+            foreach (Volunteer::where('training_session_id', $training_session_id)->get() as $volunteer) {
+                foreach (Status::all() as $status) {
+                    if ($volunteer->id == $status->volunteer_id) {
+                        Status::find($status->id)->update(['acceptance_status' => 1]);
+                    }
                 }
+            }
+            ApprovedApplicant::truncate();
         }
-        ApprovedApplicant::truncate();
-
         return redirect()->back();
     }
 
@@ -870,7 +874,7 @@ class TrainingSessionController extends Controller
 
     public function trainingCenterShow(TrainingSession $trainingSession, TraininingCenter $trainingCenter)
     {
-        $cindicationRooms = CindicationRoom::where('training_session_id',$trainingSession->id)->where('trainining_center_id',$trainingCenter->id)->get();
+        $cindicationRooms = CindicationRoom::where('training_session_id', $trainingSession->id)->where('trainining_center_id', $trainingCenter->id)->get();
         $miniSide = 'aside-minimize';
         $volunteers = Volunteer::whereRelation('approvedApplicant.trainingPlacement.trainingCenterCapacity.trainingCenter', 'id', $trainingCenter->id)->get();
         $checkedInVolunteers = Volunteer::whereRelation('approvedApplicant.trainingPlacement.trainingCenterCapacity.trainingCenter', 'id', $trainingCenter->id)->whereRelation('status', 'acceptance_status', 5)->get();
@@ -887,7 +891,7 @@ class TrainingSessionController extends Controller
         $centerCheckers = $centerCheckerQuery->get();
         Permission::findOrCreate('checker');
         $checkerUsers = User::doesntHave('volunteer')->doesntHave('trainner')->permission('checker')->whereNotIn('id', $centerCheckerQuery->pluck('id'))->get();
-        return view('training_session.center_show', compact('centerCoordinators', 'centerCoordinatorUsers', 'checkedInVolunteers', 'centerCheckers', 'checkerUsers', 'freeTrainners', 'trainings', 'trainingSession', 'totalTrainingMasters', 'totalVolunteers', 'trainingCenter', 'miniSide','cindicationRooms'));
+        return view('training_session.center_show', compact('centerCoordinators', 'centerCoordinatorUsers', 'checkedInVolunteers', 'centerCheckers', 'checkerUsers', 'freeTrainners', 'trainings', 'trainingSession', 'totalTrainingMasters', 'totalVolunteers', 'trainingCenter', 'miniSide', 'cindicationRooms'));
     }
     public function resourceAssignToTrainingCenter($training_session, Request $request)
     {
